cherry-pick ef282135935dc2df0019e439bf41f6f77d63b3c8

+++ app/Services/Assetlite/ComponentService.php
@@ -190,7 +190,31 @@ class ComponentService
 
         // }
 
-        $this->save($data);
+
+        $component = $this->save($data);
+
+        if (($data['component_type'] == "multiple_image" || $data['component_type'] == "features_component") && isset($data['base_image'])) {
+            foreach ($data['base_image'] as $key => $img) {
+                if (!empty($img)) {
+                    $baseImgUrl = $this->upload($img, 'assetlite/images/component');
+                }
+                $imgData = [
+                    'component_id' => $component->id,
+                    'page_type' => $pageType,
+                    'title_en' => isset($data['multi_title_en']) ? $data['multi_title_en'][$key] : '',
+                    'title_bn' => isset($data['multi_title_bn']) ? $data['multi_title_bn'][$key] : '',
+                    'alt_text_en' => $data['multi_alt_text_en'][$key],
+                    'alt_text_bn' => $data['multi_alt_text_bn'][$key],
+                    'img_name_en' => str_replace(' ', '-', strtolower($data['img_name_en'][$key])),
+                    'img_name_bn' => str_replace(' ', '-', strtolower($data['img_name_bn'][$key])),
+                    'base_image' => $baseImgUrl,
+                    'created_by' => Auth::id(),
+                ];
+                $this->comMultiDataRepository->save($imgData);
+            }
+        }
+
+
         return response('Component create successfully!');
     }
 
@@ -204,10 +228,10 @@ class ComponentService
             ]);
         }
 
-        request()->validate([
-            'image_name_en' => 'unique:components,image_name_en,' . $id,
-            'image_name_bn' => 'unique:components,image_name_bn,' . $id,
-        ]);
+        // request()->validate([
+        //     'image_name_en' => 'unique:components,image_name_en,' . $id,
+        //     'image_name_bn' => 'unique:components,image_name_bn,' . $id,
+        // ]);
 
         $component = $this->findOne($id);
         if (request()->hasFile('image')) {
@@ -308,7 +332,7 @@ class ComponentService
 
         $component->update($data);
 
-        if ($data['component_type'] == "multiple_image" || $data['component_type'] == "features_component") {
+        if (($data['component_type'] == "multiple_image" || $data['component_type'] == "features_component") && isset($data['base_image'])) {
             $this->comMultiDataRepository->deleteAllById($id);
             foreach ($data['base_image'] as $key => $img) {
 //                dd($data);
